requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.2
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.25.0
mdurl==0.1.2
multidict==6.7.0
numpy==2.2.6
openai==2.15.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.1.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.30.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.50.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
yarl==1.22.0
